mydjangoproject/requirements.txt
django-bootstrap4
django-bootstrap-themes
django-crispy-forms
pillow
djangorestframework
djangorestframework-jwt
requests